feat: All 4 final steps — complete cross-platform offline payment SDK
Step 1: Real QR Scanners
- Android: CameraX + ML Kit barcode scanning (QRScannerScreen.kt)
- Android: ZXing QR code generation (replaces placeholder ReceiveSheet)
- iOS: AVFoundation QRScannerView with animated viewfinder overlay + CoreBluetooth

Step 2: Rust Build Scripts
- build-android.sh: cargo-ndk → .so for arm64/armv7/x86_64 + Kotlin UniFFI bindings
- build-ios.sh: lipo fat binary → XCFramework + Swift UniFFI bindings

Step 3: Mesh Packet Persistence (Spray-and-Wait survive restarts)
- Android: Room DB (AstralDatabase.kt) with MeshPacket + Transaction entities
- iOS: CoreData (AstralPersistence.swift) with encrypted-at-rest store
- Both: 7-day TTL expiry, dedup, forward count tracking

Step 4: Swift Package Manager
- Package.swift: CoreData + CoreBluetooth + Network + AVFoundation linked
- CONTRIBUTING.md: SPM + Android integration docs

+++ astral-core/src/errors.rs
@@ -7,7 +7,7 @@ pub enum AstralError {
     #[error("Double spend detected. Transaction counter {counter} invalid")]
     DoubleSpendDetected { counter: u64 },
     #[error("Packet reassembly timed out for session {session_id}")]
-    ReassemblyTimeout { session_id: u32 },
+    ReassemblyTimeout { session_id: u16 },
     #[error("Payload exceeds maximum protocol capacity: {size} bytes")]
     PayloadTooLarge { size: u32 },
 }
